- unpacking bsas during installation no longer overwrites loose files (since the loose files would have taken precedence)

+++ src/bsaarchive.cpp
@@ -32,6 +32,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 #include <boost/interprocess/sync/scoped_lock.hpp>
 #include <boost/interprocess/sync/interprocess_semaphore.hpp>
 #include <zlib.h>
+#include <sys/stat.h>
 #define WIN32_LEAN_AND_MEAN
 #include <Windows.h>
 
@@ -519,11 +520,19 @@ void Archive::readFiles(std::queue<FileInfo> &queue, boost::mutex &mutex,
 }
 
 
+inline bool fileExists(const std::string &name) {
+  struct stat buffer;
+  return stat(name.c_str(), &buffer) != -1;
+}
+
+
 void Archive::extractFiles(const std::string &targetDirectory,
                            std::queue<FileInfo> &queue, boost::mutex &mutex,
                            boost::interprocess::interprocess_semaphore &bufferCount,
                            boost::interprocess::interprocess_semaphore &queueFree,
-                           int totalFiles, int &filesDone)
+                           int totalFiles,
+                           bool overwrite,
+                           int &filesDone)
 {
   for (int i = 0; i < totalFiles; ++i) {
     bufferCount.wait();
@@ -544,6 +553,10 @@ void Archive::extractFiles(const std::string &targetDirectory,
     DataBuffer dataBuffer = fileInfo.data;
 
     std::string fileName = makeString("%s\\%s", targetDirectory.c_str(), fileInfo.file->getFilePath().c_str());
+    if (!overwrite && fileExists(fileName)) {
+      continue;
+    }
+
     std::ofstream outputFile(fileName.c_str(), fstream::out | fstream::binary | fstream::trunc);
 
     if (compressed(fileInfo.file)) {
@@ -557,7 +570,7 @@ void Archive::extractFiles(const std::string &targetDirectory,
       BSAULong length = 0UL;
       try {
         boost::shared_array<unsigned char> buffer = decompress(dataBuffer.first.get(), dataBuffer.second + sizeof(BSAULong),
-                                                                     result, length);
+                                                               result, length);
         if (buffer.get() != NULL) {
           outputFile.write(reinterpret_cast<char*>(buffer.get()), length);
         }
@@ -583,7 +596,9 @@ void Archive::createFolders(const std::string &targetDirectory, Folder::Ptr fold
 }
 
 
-EErrorCode Archive::extractAll(const char *outputDirectory, const boost::function<bool (int value, std::string fileName)> &progress)
+EErrorCode Archive::extractAll(const char *outputDirectory,
+                               const boost::function<bool (int value, std::string fileName)> &progress,
+                               bool overwrite)
 {
 #pragma message("report errors")
   createFolders(outputDirectory, m_RootFolder);
@@ -608,7 +623,7 @@ EErrorCode Archive::extractAll(const char *outputDirectory, const boost::functio
                               outputDirectory,
                               boost::ref(buffers), boost::ref(queueMutex),
                               boost::ref(bufferCount), boost::ref(queueFree),
-                              fileList.size(), boost::ref(filesDone)));
+                              fileList.size(), overwrite, boost::ref(filesDone)));
 
   bool readerDone  = false;
   bool extractDone = false;
